Version 3.0.0
- Attack damage and attack speed parameters for enhanced tool variants and EnhancedToolTier#createEnhancedToolAttributes are now doubles (previously integers)
  - String parameter for the resource location is removed. The resource location path is defined within the method.

- Added TagKey registry for TagKeys

- Method for creating additional recipes is now called before other recipe creation methods

+++ src/main/java/net/laserdiamond/laserutils/item/equipment/tools/EnhancedAxeItem.java
@@ -15,7 +15,7 @@ public class EnhancedAxeItem extends AxeItem {
      * @param attackSpeed The attack speed
      * @param properties The {@link Properties} to give the {@link EnhancedAxeItem}
      */
-    public EnhancedAxeItem(EnhancedToolTier enhancedToolTier, float attackDamage, float attackSpeed, Properties properties) {
+    public EnhancedAxeItem(EnhancedToolTier enhancedToolTier, double attackDamage, double attackSpeed, Properties properties) {
         super(enhancedToolTier.toolTier(), properties.attributes(EnhancedToolTier.createEnhancedToolAttributes(enhancedToolTier, attackDamage, attackSpeed).build()));
     }
 }
